requirements.txt
tabula-py==2.7.0
pandas==2.1.0
telebot==0.0.5
pyTelegramBotAPI==4.10.0
requests==2.28.2
beautifulsoup4==4.12.2
camelot-py==0.9.0

